Block admin delete is admin < 2 updated dash

+++ app/Http/Controllers/management/UserManagement.php
@@ -33,12 +33,19 @@ class UserManagement extends Controller
   }
 
   public function destroy($id)
-  {
-      $user = User::findOrFail($id);
-      $user->delete();
+{
+    $usersWithAdminTwo = User::where('admin', 2)->get();
 
-      return back()->with('success', 'User has been deleted successfully!');
-  }
+    // Check if there are 2 or more users with admin=2
+    if ($usersWithAdminTwo->count() >= 2) {
+        $user = User::findOrFail($id);
+        $user->delete();
+        return back()->with('success', 'User has been deleted successfully!');
+    }
+
+    // Return back with error message
+    return back()->with('failed', 'Cannot delete the user. At least two users must be Admins.');
+}
 
   public function blockUser(Request $request, $user)
   {
